refactor: Overall refactoring of RNA Class.
Currently the class has a lot of methods that work only with precisely formatted input, which greatly reduces the usability of the class. To alleviate this, RNA class needs to be refactored to accept and handle data more dynamically, without realy being cornerned where the data comes from and how it's formatted.

The class also has many functions that are not relevant to it, but has been used to perform various tasks that are utilizing the class methods. These methods will be removed.